requirements.txt
Flask
flask-cors
flask-mysqldb
werkzeug
numpy
opencv-python
Pillow